# src/services/hotel-v1-python/app/swagger/openapi.json
{"openapi":"3.0.0","info":{"title":"Bee Travels Hotel V1 Service","version":"1.0.0","description":"replace this with a real description of your service","license":{"name":"Apache 2.0","url":"https://github.com/bee-travels/bee-travels-node/blob/master/LICENSE"}},"paths":{"/live":{"get":{"description":"Liveness check to make sure service is available","responses":{"200":{"description":"OK"}}}},"/ready":{"get":{"description":"Readiness check to make sure service and all connected external service calls are available","responses":{"200":{"description":"OK"},"503":{"description":"Service Unavailable"}}}},"/api/v1/hotels/info/{filter}":{"get":{"tags":["Hotel"],"summary":"Get filter list","description":"Gets list of a type to filter Hotel data by.","parameters":[{"name":"filter","in":"path","description":"The name of the filter to get options for.","required":true,"schema":{"$ref":"#/components/schemas/FilterType"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Filter Not Found Error"},"500":{"description":"Internal Server Error"}}}},"/api/v1/hotels/{id}":{"get":{"tags":["Hotel"],"summary":"Get hotel by id","description":"Gets data associated with a specific Hotel ID.","parameters":[{"name":"id","in":"path","description":"id of the Hotel","required":true,"schema":{"type":"string"}},{"name":"dateFrom","in":"query","description":"Date From","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","description":"Date To","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"404":{"description":"not found"},"500":{"description":"Internal Server Error"}}}},"/api/v1/hotels/{country}/{city}":{"get":{"tags":["Hotel"],"summary":"Get list of hotels","description":"Gets data associated with a specific city.","parameters":[{"name":"country","in":"path","description":"Country of the hotel using slug casing.","required":true,"schema":{"type":"string"}},{"name":"city","in":"path","description":"City of the hotel using slug casing.","required":true,"schema":{"type":"string"}},{"name":"dateFrom","in":"query","description":"Date From","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","description":"Date To","required":true,"schema":{"type":"string"}},{"name":"superchain","in":"query","description":"Hotel superchain name.","required":false,"schema":{"type":"string"}},{"name":"hotel","in":"query","description":"Hotel Name.","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","description":"Hotel Type.","required":false,"schema":{"type":"string"}},{"name":"mincost","in":"query","description":"Min Cost.","required":false,"schema":{"type":"number"}},{"name":"maxcost","in":"query","description":"Max Cost.","required":false,"schema":{"type":"number"}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"FilterType":{"type":"string","enum":["superchain","name","type"]}}}}